Initial commit.
[[This repository was converted from CVS to Git on 2013-12-26 by
Mike Swanson. Junk commits are attempted to be removed, tags
renamed into a more sensible manner, but for the most part, the
repository remains intact of its historical state.

ZSNES is ancient and dead software, nobody should use it anymore,
and because this is not a repository intended to continue active
development, it may be overwritten entirely (push --force) in the
future. Use at own risk.]]